Removed most of Working with People
This is the first in a series of commits designed to streamline the
exercise for tech teams getting started with chef/configuration
management/automation. The Working with People section can enable
quicker and/or smother starts, but is more about what organizations need
to do to sustain improvement. As such, it is not needed for a new team.