Removed several Continuous Integration lines
+ The CI service automatically tests new branches
+ CI confirms that versions are unique
+ CI jobs automatically update dependencies
All three of these are belong in a mature CI implementation, but are not
essential to getting started.

+ CI makes the quality of the code base highly visible
This is in place so that Justin (me) can rant about CI jobs that detect
errors that don't break the build. We don't need an obtuse statement for
this.

+ CI jobs use monitoring to assess each change's effect on system health
This is an adavnced state. Folks who are going to implement this, are
not going to do this exercise.