Removed two Virtualization as a Service lines
+ The API is a generally accepted API such as EC2, Azure, OpenStack,
vSphere, Docker
Folks using a homegrown API are a corner case. We just don't see it
enough to merit talking about it with everyone.

+ Access restrictions allow authorized users (e.g., developers and
operations) to provision resources, and deny unauthorized users
The folks who have problems with this know they have problems and are
not looking for help from us with it.